add default 1 category and channel creation
